Clear out test-only code
