+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/PlatformSpecifics_iOSTranslucentNavBar.cs
@@ -1,6 +1,5 @@
 using Xamarin.Forms.CustomAttributes;
 using Xamarin.Forms.Internals;
-using ImAVendor.Forms.PlatformConfiguration.iOS;
 using Xamarin.Forms.PlatformConfiguration.iOS;
 
 namespace Xamarin.Forms.Controls
@@ -16,7 +15,7 @@ namespace Xamarin.Forms.Controls
 			var button = new Button { Text = "Toggle Translucent", BackgroundColor = Color.Yellow };
 
 			button.Clicked += (sender, args) => On<iOSPlatform>()
-				.EnableTranslucentNavigationBarVendor(!On<iOSPlatform>().IsNavigationBarTranslucentVendor());
+				.EnableTranslucentNavigationBar(!On<iOSPlatform>().IsNavigationBarTranslucent());
 
 			var content = new ContentPage
 			{
