+++ Xamarin.Forms.Xaml.UnitTests/PlatformSpecifics.xaml
@@ -2,9 +2,9 @@
 <forms:MasterDetailPage xmlns="clr-namespace:Xamarin.Forms.Xaml.UnitTests;assembly=Xamarin.Forms.Xaml.UnitTests"
 			 xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
 			 xmlns:forms="http://xamarin.com/schemas/2014/forms"
-			 xmlns:android="clr-namespace:Xamarin.Forms.PlatformConfiguration.Android;assembly=Xamarin.Forms.Core"
+			 xmlns:win="clr-namespace:Xamarin.Forms.PlatformConfiguration.Windows;assembly=Xamarin.Forms.Core"
 			 x:Class="Xamarin.Forms.Xaml.UnitTests.PlatformSpecific"
-			 android:MasterDetailPage.SomeAndroidThing="9999"
+			 win:MasterDetailPage.CollapseStyle="Partial"
 			 x:Name="TestMDP">
 	<forms:MasterDetailPage.Master>
 		<forms:ContentPage x:Name="masterPage" Title="Platform Specifics" />
